New execute_processing_block Celery task
This is a 'drop-in' replacement for sip_pbc's execute_processing_block
task. Instead of working with the idea of workflows, etc, it simply
takes the processing block and turns is into a DALiuGE graph, which is
then executed on a normal DIM/NM setup.

As previously, regardless of the workflow indicated by the
ProcessingBlock, we are always executing the same graph, which we
internally also turn into a "zerorun" graph -- one where all
applications are turned into a "SleepApp" with zero sleeping time.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>