New main module to be invoked in the K8S container
It's basically a copy/adjustment of those existing in the SDP
proptotype. It gets the SB that it's supposed to run, prepares a daliuge
deployment (with a Helm chart that doesn't exist yet), and submits a
logical graph for execution into the NMs that will be running in the K8S
infrastructure after the chart is created.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>